# Fixed latency setting and cleaned-up OpenAL backend