--- a/Source/Core/AudioCommon/OpenALStream.cpp
+++ b/Source/Core/AudioCommon/OpenALStream.cpp
@@ -290,7 +290,7 @@ void OpenALStream::SoundLoop()
     palGetSourcei(m_source, AL_BUFFERS_PROCESSED, &num_buffers_processed);
     if (num_buffers_queued == OAL_BUFFERS && !num_buffers_processed)
     {
-      m_sound_sync_event.Wait();
+      std::this_thread::sleep_for(std::chrono::milliseconds(1));
       continue;
     }