--- a/Source/Core/Core/ConfigManager.h
+++ b/Source/Core/Core/ConfigManager.h
@@ -109,7 +109,7 @@ struct SConfig : NonCopyable
   bool bCopyWiiSaveNetplay = true;
 
   bool bDPL2Decoder = false;
-  int iLatency = 14;
+  int iLatency = 20;
   bool m_audio_stretch = false;
   int m_audio_stretch_max_latency = 80;