Before these changes each unit of latency was actually 5ms, with a minimum latency of ~10 ms. If it was set to 4 ms on the UI, the actual latency was 10 ms + 5 ms (for each buffer, 256 samples at 48kHz) * 4 ms = 30 ms.
Now 30 ms on the UI means 30 ms on the backend.

~~**Problem:** People that had "low" values on their latency setting may think that the latency is now higher (even though it can be lower as I also commented out a Wait() event that did more harm than good) and they will need to increase their values to 20 or 30 ms.~~
I changed the setting in the INI from "Latency" to "AudioLatency" so everyone should have a good default.
Instead of commenting out the Wait() event (which I soon discovered caused high CPU usage, duh), I swapped it out for a std::sleep_for. Works a lot better now, only causes buffer underruns when the latency is set too low.

I also renamed all variables to the current coding standard, renamed some variables to better represent what they really are, removed a redundant conversion from short to floating point when playing back stereo and swapped all C-style arrays to std::array.

This PR is partially based on PR# 5191 and PR# 5235. Some changes I believe are particularly important (such as the removal of the redundant conversion from short to float) so I want to move it from PR# 5235 which will only be merged when I manage to discover why FreeSurround causes glitches on Ubuntu 16.04.

EDIT:

Just did some experiments with this PR and OpenAL Soft. If we set the latency to 15 ms and reduce the period size in its config file to 128 we get the same audio latency as XAudio2 😊

I also investigated why commenting out the Wait() event allows lower latencies.
After we call Wait() the thread unpauses only when SoundStream.Update() is called and samples are pushed into the mixer by SendAIBuffer(), but for some reason this doesn't work very well with OpenAL and fixed sized buffers. I made some changes on another branch where OpenAL receives and buffers all available samples at once hoping to lower its latency. The latency continued the same but works really well with Wait().

Luckily as OpenAL is in sync with the mixer (as it plays samples at the same rate it receives them) changing Wait() to sleep_for brings only advantages.