@JMC47 2017-06-17
I've ran into at least 2 people who've had issues with the new OpenAL latency, so I'd be happy with anything that makes it easier to use.