@MerryMage 2017-06-17
Two comments:

* Add "ms" to the GUI after the latency textbook.
* Perhaps rename the INI setting from "Latency" to "AudioLatency" due to the change in semantics of the given value.