@LAGonauta 2017-06-17
@MerryMage Good idea, now everyone gets a good default for a start.