@ligfx 2017-06-18
Looks good to me. Thanks for fixing the unit-less "Latency" option, that's bugged me for so long!

Could we lower the default latency? 30ms seems high to me, other backends are between about 5ms to 10ms.