@LAGonauta 2017-06-18
Just reduced the default from 30 ms to 20 ms (latency between 14 and 20 ms). I could reduce it to 10ms (latency between 7 and 10 ms), but it only works well if we edit OpenAL Soft's configuration file and set period_size to 128 samples (the default is 1024).
Of course with a X-Fi we can set it to 6ms and it not only works fine (latency between 4 to 6 ms).