@LAGonauta 2017-06-27
Just rebased this PR with the OpenAL changes by ligfx